grunt: Introduce autoloader task
This is our first foray into Grunt, and seems to be shaping up nicely.
This task is needed to generate the fallback files utilized by
WordPoints’s autoloader when SPL autoloading isn’t available. It was
originally included in the hooks API module’s grunt config file, and
was merged into WordPoints core when the initial version of the hooks
API was merged in. We now extract the basic code into this task that
can be reused across WordPoints projects.

While we were at it, we also added a check that the generated file
actually executes correctly.

Since we are now including JS files in the dev-lib, we added a JSHint
check to our Travis build. We currently configure JSHint exactly as we
do anything else, by symlinking the included config file to our root
directory.

Fixes #126