grunt: loop over all autoloader directories
Instead of hardcoding just a single one. We also now automatically
create the autoloader fallback file in a directory of classes if it
doesn’t already exist.

One result of these changes was that we realized that classes in one
directory may depend upon classes in another directory. This meant that
the autoloader files would fail to validate even if they would actually
work in real life. To solve this, we’ve updated our verification code
to load the autoload files that the checked file might depend upon when
verifying it.

See #126