phpcs: report missing long condition comments
Instead of blocking this completely (because of issues with the format it demands for end comments) we can just silence reports of end comments with the incorrect format. This allows us to continue to get reports of new conditions that are too long, but to use our own format for the comments once we add them.

See #181