travis: Support for network wpcept tests
Tests that need to be run with WordPoints network active can now be placed in the `network` directory within the `acceptance` Codeception test suite. We don't run this for modules yet, it is still experimental and we don't have formal commands for it yet, just a basic one that can be used with custom arguments and env vars.

In the process it was discovered that we are using array expansion like `${@:1}`, but the `:1` part is unnecessary, since expansion always starts from the first element in the array. To actually skip the first element, we'd need to use `:2`. But we don't need to do that. We were just assuming that this started from `:0`, but it doesn't.

See #189