phpunit: copy bootstrap from WordPoints
All of the version numbers are updated to be 2.6.0, the version of the dev-lib that they will first appear in, and the package has been updated to `WordPoints_Dev_Lib` instead of `WordPoints`. The existing dev-lib PHPUnit classes have been moved to the `classes-deprecated` directory, but soon we'll introduce new versions of them lacking the `Dev_Lib` prefix that will live in the `classes` directory.

The bootstrap loader class is slightly modified to load the remote module/component install files from the correct locations.

The WordPoints bootstrap has also been modified slightly, with legacy handling for old versions of the dev-lib removed.

See WordPoints/wordpoints#587, #193