git: add pre-commit hook
The added and modified files that are staged for commit will automatically be checked. Note that we will check against the actual working copy of the files, not the staged changes, which in some cases could in theory be different. But right now we don't care.

What we do in the git pre-commit hook for now is just run all of the codesniff tools. We have updated the `wpdl-codesniff` function to return `1` when there are any codesniff errors, whereas before it would just return the return code of the last command.

When the pre-commit hook is run, the checks are limited to only the staged files, as mentioned above. This is done via the introduction of the `wpdl-get-codesniff-files` command, which finds all of the files that would normally be sniffed by a particular tool, and then filters them to be only those that are also staged for commit. The list of files is output as a `\0`-delimited string, for better handling with `xargs`, which many of the commands are now updated to use. On that note, we were doing some things all wrong before in how we were passing the list of found files around, but this is now very much improved with the use of `\0` delimited strings and `xargs -0`.

The various codesniffing command have been updated to use this new method of getting the list of files, with the exceptions of the autoloaders checker, the symlinks checker, and the `wp-l10-validator`. For the first two this is just because they are more complex to restrict, and I just haven't tackled them yet. The l10n validator, on the other hand, doesn't actually support being run against an individual file(s), so we'll either have to change that upstream or create our own command that utilizes it.

While working on this I realized that we aren't ignoring the `.git` directory when codesniffing, which is silly. I've moved the ignore rule for that from the strings sniff path to the generic codesniff rules. I've also removed the `-type f` from the strings path definition, because it didn't appear to work, and is redundant now that we always do that in the file getter command anyway.

Fixes #20