phpunit: fail gracefully when missing targets
Previously we were calling `fail()` outside of a test method, which just killed the tests completely. And on newer versions of PHPUnit, this doesn't even exit with an error code, leading to builds not properly failing on Travis CI.

I also added `array_unique()` on the reported values, since otherwise the target will be repeated for every reactor that it was tested against.