phpunit: support requiring WordPoints versions
Adds support for the `@requires WordPoints version` and `@WordPoints-version <version>` annotations to require a WordPoints version for a unit test. The reason that we use two annotations is so that we can check for the requirement with `isset()`, since the dynamic part (version number) is provided separately via the other annotation.

We wanted to use `@WordPoints\version` instead of `@WordPoints-version`, but PHPUnit doesn't grab annotations that contain backslashes.

We'll use this in relation to WordPoints/buddypress#7