makepot: fix error for max_header_lines
The `$max_header_lines` property is currently marked as `private` in the parent class. Until that is fixed, we need to define it in the child class to work around errors like this:

```
PHP Notice:  Undefined property: WordPoints_MakePOT::$max_header_lines in /Users/johngrimes/git/_wordpoints-modules/democracy-poll/dev-lib/i18n/makepot.php on line 117
PHP Stack trace:
PHP   1. {main}() /Users/johngrimes/git/_wordpoints-modules/democracy-poll/dev-lib/i18n/makepot.php:0
PHP   2. WordPoints_MakePOT->wordpoints_module() /Users/johngrimes/git/_wordpoints-modules/democracy-poll/dev-lib/i18n/makepot.php:212
```

The result is that the command fails, because the number of lines is then interpreted as 0.

The default for this value is 30 in the parent class, but because our "headers" are wrapped in a function, we have a header comment in front of them, which usually puts them just over the limit.

See #145