phpcs: Stop ignoring warnings
Previously, when running PHPCS from the command line we have been passing thr `-n` option, which causes only error-level messages to be reported, and not warnings. As time has passed, it has occurred to me that this is really not necessary—one reason that we did it originally is because some VIP sniffs produced a bunch of warnings that we did not care about—and as things evolve we'll be missing out on some things. Actually, we've been missing out on some things that were converted from errors to warnings or have just been warnings all along.

We've updated our ruleset to take this change into account, by silencing any warnings that we aren't interested in.

- We use underscores in dynamic hook names, see https://github.com/WordPress-Coding-Standards/WordPress-Coding-Standards/issues/751#issuecomment-278739935.
- The `customVariablesWhitelist` property is deprecated.
- Some file system function that we use in some (deprecated) test helper functions.
- The `// }` in the test class maps that marks where the dynamically generated portion ends is being flagged as commented-out code.
- ASP-style tags used in our old Backbone templates for the ranks screen, which we'll probably be deprecating before long.

See WordPoints/wordpoints#606