phpunit: Alternatives to assertContains()
The `assertContains()` method does non-strict comparisons by default. We want to avoid this, so we are going to outlaw the method. To allow us to continue to easily replace it though, we need to take into account the two different pieces of functionality it offers: searching in a string, and searching in an array. For the first, we introduce `assertStringContains()`, and for the latter, `assertContainsSame()`.

See WordPoints/wordpoints#567