autoloader: Allow interface files to be filtered
We allow the class files to be filtered mainly so that the order can be re-arranged, since some classes extend others. The same can actually be true of interfaces, and we just ran into an issue with that, so it seems that we should also add the interface files before running the filter, instead of after.