l10n-validator: fix checking for changed files
When running the git precommit checks, our check of the staged files
against the list of files that were in scope would always return no
matches, because the paths were prefixed with just `src/` instead of
`./src/` as expected.

This meant that the l10n validator never ran during precommit.

See #203