phpunit: don't deactivate plugin if only uninstalling module
When only the module is being uninstalled, but not WordPoints itself,
the module should be deactivated alone, rather than the entire plugin
being deactivated.

The plugin deactivation is automatically being done by the plugin usage
simulation file that we are using to load the module usage simulation
file, so to skip the plugin deactivation, we just exit early.