wpcept: avoid errors on update screen
The list of updates is assumed to be an array, and if it is not it will
result in `list_core_updates()` being called with `false` instead of the
data for an update, which will cause errors. An empty array as the
updates property ensures that WordPress will correctly detect that no
core updates are available.