wpcept: Fix have module with update method
Due to changes in WordPoints, the method would no longer work, because
the data it saved for a module would end up getting overwritten. First,
it would be overwritten by an update check, because the list of modules
checked could be different. And second, it would be overwritten when a
license check was performed. We now avoid the license check by marking
the module as free.

See WordPoints/wordpoints#645