phpunit: provide createMock() and createPartialMock()
These methods were introduced in newer versions of PHPUnit, to replace
the `getMock()` method, which was deprecated and removed in PHPUnit 6.

Since we have to run on both old versions that don't have these
functions yet, as well as newer versions where `getMock()` is
deprecated, we need to "backport" these newer methods.

Fixes #222