wpcept: set the value of the server API for extensions
We are experiencing failures on Travis CI for tests that use this
method, and I suspect it is because Cloudflare is blocking the requests.
It isn't important that the tests actually make that request anyway; in
fact, it would be good to avoid it as it speeds things up. We do that by
setting the transient value ahead of time.