phpunit: Avoid errors when testing extension only uninstall
Because WordPoints itself isn't installed, it is fully loaded, and thus
all hooks are hooked up. Including some that perform actions when a blog
is deleted. These hooks end up calling `is_wordpoints_network_active()`,
which is not supposed to be used during uninstall. So when we delete the
site used in the tests, that error will be thrown. We avoid this by
filtering the result of the `wordpoints_is_uninstalling()` function.

In order to make this change, we needed to update to WPPPB 0.3.0. We
also incorporated some of the changes from WPPPB into our own code that
is based on it.

We also make sure that WPPPB is updated when the dev-lib is updated.