phpunit: properly handle extension updates when network active
Extensions that use the new installables API were not being updated
correctly, because it was always assumed that they were network active
when WordPoints was network active. This is not the case. In fact, the
extensions are not network active by default, and so update tests for
them would fail.

See WordPoints/wordpoints#404