strings: sniff for deprecated button classes
The `button-secondary` class is deprecated, and the `button-primary`
class should always be accompanied by `button`.

While we are at it, we format the command to be more organized and
readable, and also add `--color=auto` for easier to read results.

See WordPoints/wordpoints#566

Widgets: remove inline styles for widget errors

They are nice, but just aren't a necessity, and moving them to a new
file didn't seem justified. So for now we're just removing them.

See #582